Extract klog_daemon lifecycle and shutdown orchestration
