+++ src/kernel/klog_daemon/src/main.rs
@@ -1,7 +1,8 @@
 mod cluster;
 mod config;
+mod lifecycle;
 
-use cluster::{initialize_cluster_if_needed, spawn_auto_join_task, stop_auto_join_task};
+use cluster::{initialize_cluster_if_needed, spawn_auto_join_task};
 use config::KLogRuntimeConfig;
 use klog::logs::SqliteLogStorage;
 use klog::network::{KNetworkFactory, KNetworkServer};
@@ -9,6 +10,7 @@ use klog::state_machine::{KLogStateMachine, SnapshotManager};
 use klog::state_store::{
     KLogStateStore, KLogStateStoreManager, RocksDbSnapshotMode, RocksDbStateStore,
 };
+use lifecycle::run_server_lifecycle;
 use log::{error, info};
 use openraft::Config;
 use simplelog::{ColorChoice, Config as SimpleLogConfig, LevelFilter, TermLogger, TerminalMode};
@@ -145,41 +147,5 @@ async fn run(cfg: KLogRuntimeConfig) -> Result<(), String> {
 
     let network_server = KNetworkServer::new(cfg.listen_addr.clone(), raft);
     info!("Starting raft RPC server: listen_addr={}", cfg.listen_addr);
-    let server_result = network_server.run_with_shutdown(shutdown_signal()).await;
-
-    stop_auto_join_task(join_task).await;
-
-    server_result
-}
-
-async fn shutdown_signal() {
-    let ctrl_c = async {
-        if let Err(e) = tokio::signal::ctrl_c().await {
-            error!("Failed to listen for ctrl-c shutdown signal: {}", e);
-        } else {
-            info!("Received ctrl-c shutdown signal");
-        }
-    };
-
-    #[cfg(unix)]
-    let terminate = async {
-        match tokio::signal::unix::signal(tokio::signal::unix::SignalKind::terminate()) {
-            Ok(mut sigterm) => {
-                sigterm.recv().await;
-                info!("Received SIGTERM shutdown signal");
-            }
-            Err(e) => {
-                error!("Failed to listen for SIGTERM shutdown signal: {}", e);
-                std::future::pending::<()>().await;
-            }
-        }
-    };
-
-    #[cfg(not(unix))]
-    let terminate = std::future::pending::<()>();
-
-    tokio::select! {
-        _ = ctrl_c => {}
-        _ = terminate => {}
-    }
+    run_server_lifecycle(network_server, join_task).await
 }
